+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+- Added `Element`/`AsyncElement.parent` property.
+
 ## [0.0.9]
 
 - Fixed a regression crash during script evaluation caused by missing